+++ libclamav/png.c
@@ -1,13 +1,12 @@
 /*
- *   Copyright 1995-2007 by Alexander Lehmann <lehmann@usa.net>,
- *                          Andreas Dilger <adilger@enel.ucalgary.ca>,
- *                          Glenn Randers-Pehrson <randeg@alum.rpi.edu>,
- *                          Greg Roelofs <newt@pobox.com>,
- *                          John Bowler <jbowler@acm.org>,
- *                          Tom Lane <tgl@sss.pgh.pa.us>
- *   Copyright (C) 2015 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
- *   Copyright (C) 2011 Sourcefire, Inc.
- *   Maintainer: Tomasz Kojm <tkojm@clamav.net>
+ *   Copyright (C) 2013-2019 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
+ *   Copyright (C) 2011-2013 Sourcefire, Inc.
+ *   Copyright (C) 1995-2007 by Alexander Lehmann <lehmann@usa.net>,
+ *                              Andreas Dilger <adilger@enel.ucalgary.ca>,
+ *                              Glenn Randers-Pehrson <randeg@alum.rpi.edu>,
+ *                              Greg Roelofs <newt@pobox.com>,
+ *                              John Bowler <jbowler@acm.org>,
+ *                              Tom Lane <tgl@sss.pgh.pa.us>\
  *
  *   Permission to use, copy, modify, and distribute this software and its
  *   documentation for any purpose and without fee is hereby granted, provided
@@ -710,7 +709,7 @@ int cli_parsepng(cli_ctx *ctx)
             } else if (cur_y >= h) {
                 inflateEnd(&zstrm);
 		if(eod - p > 0) {
-		    cli_dbgmsg("PNG:  %d bytes remaining in buffer before inflateEnd()", eod-p);
+		    cli_dbgmsg("PNG:  %u bytes remaining in buffer before inflateEnd()", (unsigned int)(eod - p));
 		    return CL_EPARSE;
 		}
 		err = Z_STREAM_END;
