+++ unit_tests/check_jsnorm.c
@@ -1,8 +1,8 @@
 /*
  *  Unit tests for JS normalizer.
  *
- *  Copyright (C) 2015 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
- *  Copyright (C) 2008 Sourcefire, Inc.
+ *  Copyright (C) 2013-2019 Cisco Systems, Inc. and/or its affiliates. All rights reserved.
+ *  Copyright (C) 2008-2013 Sourcefire, Inc.
  *
  *  Authors: Török Edvin
  *
@@ -145,7 +145,7 @@ END_TEST
 
 START_TEST (test_token_dval)
 {
-	int val = 0.12345;
+	double val = 0.12345;
 	yystype tok;
 	memset(&tok, 0, sizeof(tok));
